stage 4 done. Move data saving/prcessing to DB. Spring JPA usage

+++ WebQuizEngine/src/main/java/engine/api/QuizNotFoundException.java
@@ -1,7 +1,7 @@
 package engine.api;
 
 public class QuizNotFoundException extends RuntimeException {
-    public QuizNotFoundException(int id) {
+    public QuizNotFoundException(Long id) {
         super("Quiz not found: " + id);
     }
 }
